% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_calculator.R
\name{se_calculator}
\alias{se_calculator}
\title{Calculate the standard error
This function calculates standard error provided you have the sample standard deviation and sample size}
\usage{
se_calculator(sd, n)
}
\arguments{
\item{sd}{A float or numerical e.g 2, 3.4 etc}

\item{n}{the sample or population size}
}
\value{
se object
}
\description{
Calculate the standard error
This function calculates standard error provided you have the sample standard deviation and sample size
}
\examples{
se_calculator(3,18)
}
