% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_difference_calculator.R
\name{se_diff_calculator}
\alias{se_diff_calculator}
\title{# Calculate the se of a difference provided two standard deviations and two sample sizes}
\usage{
se_diff_calculator(sd1, sd2, n1, n2)
}
\arguments{
\item{sd1}{standard deviation of the first group}

\item{sd2}{standard deviation of the second group}

\item{n1}{sample size of the first group}

\item{n2}{sample size of the second group}
}
\value{
se object - a standard error of the difference
}
\description{
# Calculate the se of a difference provided two standard deviations and two sample sizes
}
\examples{
se_diff_calculator(2,2,6,10)
}
