% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_of_MD_betweengroups_calculator.R
\name{SE_of_MD_betweengroups_calculator}
\alias{SE_of_MD_betweengroups_calculator}
\title{Title}
\usage{
SE_of_MD_betweengroups_calculator(n1, n2, sd1, sd2)
}
\arguments{
\item{n1}{sample size of group1}

\item{n2}{sample size of group2}

\item{sd1}{standard deviation of group1}

\item{sd2}{standard deviation of group2}
}
\value{
SE of MD between groups
}
\description{
Title
}
\examples{
SE_of_MD_betweengroups_calculator(n1 = 100,n2 = 100,sd1 = 4,sd2 = 3)
}
